refactor(geode-core): group postproc + interop under directory modules
Relocate post-processing and external-interop code out of the crate root
into directory-backed `postproc` and `interop` modules (pure refactor; all
relocations are git mv, every cfg gate and visibility preserved):

- src/ntff.rs    -> src/postproc/ntff.rs  (postproc::ntff)
- src/viz_vtu.rs -> src/postproc/viz.rs   (postproc::viz)
- src/palace.rs  -> src/interop/palace.rs (interop::palace)

New postproc/mod.rs and interop/mod.rs carry group-level //! rustdoc and
re-export only the submodules (no item hoisting), matching driven/analytic.

lib.rs: drop `pub mod ntff/viz_vtu/palace`; add `pub mod postproc/interop`.
The sole pre-existing flat-root re-export (the 8 ntff items) becomes a
`#[deprecated]` shim of postproc::ntff. Per the subset rule, viz/palace gain
no new flat-root shims — their only callers used module paths and are migrated.

Migrate compile-hard module-path callers (Class A) to canonical paths in
3 examples + 4 tests, and the two Class B intra-doc links in
driven/scattering.rs (crate::ntff -> crate::postproc::ntff). prelude gains
canonical postproc/interop entries. Class C/D flat-root item uses and prose
mentions are deferred to #387.
