+++ crates/geode-core/tests/palace_ingestion.rs
@@ -15,7 +15,7 @@
 
 use std::path::PathBuf;
 
-use geode_core::palace::{PALACE_DEFAULT_PORT_OHM, PalaceOracleSlot, PalaceResults};
+use geode_core::interop::palace::{PALACE_DEFAULT_PORT_OHM, PalaceOracleSlot, PalaceResults};
 
 fn fixtures_dir() -> PathBuf {
     PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("tests/fixtures/palace")
