+++ crates/geode-core/tests/patch_antenna_extraction.rs
@@ -468,7 +468,7 @@ fn benchmark_fixture_reference_point_matches_committed_results() {
 /// `results.toml` slot.
 #[test]
 fn fem_vs_palace_oracle_within_band_or_skip_with_note() {
-    use geode_core::palace::PalaceOracleSlot;
+    use geode_core::interop::palace::PalaceOracleSlot;
 
     let path = repo_root().join("benchmarks/patch_antenna/results.toml");
     let raw = fs::read_to_string(&path)
